requirements.txt
requests
pydantic-ai
pydantic-ai-slim[groq]
streamlit
python-dotenv